Do not embed unneeded json metadata in applets
We already have it for the KPackage and would otherwise parse it twice.
This also avoids a misleading warning about specifying a plugin Id for a plugin with embedded JSON metadata in KF6.

Also, use kcoreaddons_add_plugin since the target needs to be renamed to keep the derived ID the same.