snap: update to latest release
The sound for the snap doesn't work. Rest of the app works perfectly. The kclockd runs in the background, once the clock app itself is opened. I guess the audio not working is due to the qtmultimedia of snap not having any such backend. I have created a PR there also, hopefully it'll be fixed. 

https://invent.kde.org/neon/snap-packaging/kde-qt6-core-sdk/-/merge_requests/3

EDIT: is kclock using mpris for alarm sound? I guess it's fixable then.