AlarmForm: create duration and snooze models correctly
Currently these are created imperatively to work around the fact that
ListItem can't accept i18n() strings. This imperative assignment breaks
implicitWidth bindings on the Repeater object, which bubbles up and
eventually causes the dialog to have a content size of zero.

Instead, create the data models as arrays of dictionaries, which
preserves the bindings and is the recommended way to do this sort of
lightweight custom model.