Add WindowExposure attached property
Wayland has no such concept as a minimized window.
The window is "suspended" when hidden (minimized, different virtual desktop, etc)
which Qt maps to the exposed state.

Unfortunately, isExposed is not exposed (heh) to QtQuick Window API.
Hence, a WindowExposure attached property is added that is attached to
a QWindow and reports its exposure state.

This can then be used to show time information in the window title
when the window is hidden.