StopWatchTimer: Use QElapsedTimer for tracking stopwatch
It's a monotonic timer that also doesn't skew as the system time is changed.
It is also a lot more light-weight than doing proper date time calculations.

Use the QElapsedTimer state for whether it's running or not and keep track
of the last "paused" time since QElapsedTimer can only be invalidated but
not "resumed".