kclockd: Disable desktop settings awareness
This prevents loading the platform integration plug-in and saves
a substantial amount of RAM. The daemon doesn't show any UI on its own
but plays sounds and uses KNotification.

The only "UI" (and the reason it's kept as QApplication) is the QMenu
created internally by KStatusNotifierItem but that menu is typically
rendered by the tray implementation through DBus menu anyway.