StopwatchPage: Improve layout
Don't hardcode button sizes, this doesn't work with all localizations.
Put them in a container of the desired size so they are layouted nicely
but let the button take whatever size it needs.

Further, always consider lap timer when sizing the time labels.
Avoid shifting the buttons when the first lap is added, instead animate
it into view nicely.