Add persistent resident notification when timer is running
This allows to control a timer (pause, resume, reset, add a minute)
even when KClock isn't running. Particularly when launching a timer
through KRunner there was previously no indication that something
was going on.

An option is added to show the notification always, only when
KClock isn't running (the default), never.

GlobalEnums is added to kcfgc so the enum becomes available
via Q_ENUM and is now also compiled into kclock, so the same
enum can be shared between app and daemon.