Use more light-weight DBus call for starting kclockd
Never ever use QDBusInterface. It causes a blocking Introspect call
in its constructor. We then also call Introspect ourself. Introspect
is relatively heavy as the callee has to create an XML description
of its interface.

Instead, create an ad-hoc QDBusMessage calling the "Ping" method.
Result is the same, but it's a lot lighter.