Port to declarative type registration
The URI clashes with the binary name, so it's changed to typical RDN.
This also folds the "components" folder into the same import.

Also, object ownership of singletons is clarified. By default, the QML
engine would adopt them which is not what we want. Create the engine
on the stack, too.

Further, make constructors private for the singletons. It's been done
in most cases but more importantly, QML_SINGLETON calls the custom
"create" function only if there is no public default(ed) constructor
available. It will silently just create a new instance otherwise!