Implement picture-in-picture mode
Making use of xx-pip-v1 Wayland protocol.

This allows to "pop out" the stopwatch page or a timer, similar to
how it's possible with a video in a web browser. Clicking "return
to app" returns back to the page that was popped out.

PipHandler singleton manages the PIP contents. It contains a
StackView that anyone can push components into to display as
PIP window, which also ensures the PIP content stays alive after
leaving the page.

Co-Authored-By: Vlad Zahorodnii <vlad.zahorodnii@kde.org>
Co-Authored-By: David Edmundson <kde@davidedmundson.co.uk>