Add time completed information
solves #95 

Adds time completed information into the timer page. The time to completion will be updated when the user change how long the timer is and when starting the timer. This means that if you pause the timer will the time when it will be completed be stuck until you unpause. This will also show even if say you have a 5 sec timer, is this something that we might want to change?

Preview images:

![image](/uploads/645cb35cfa699a61e41dd85134f51e29/image.png){width=473 height=448}
![image](/uploads/99109c22d2de236f117051900e21d3f7/image.png){width=473 height=448}