New Alarm: Improve inputs tab order control
With this change, when the input focus is in the hours input field
and the user presses tab, focus directly goes to the minutes input
field. (Before, it jumped to the back button for some reason).

As a side-effect, the up/down buttons for hours/minutes are now
not reachable via tab navigation anymore. Not sure if this could
cause other issues.