Rework loading of alarm and timer lists
Make it asynchronous and introduce a status enum. Rather than checking
whether the daemon is running, just call into it and fail if it isn't there.
Also show the DBus error message, if applicable.

DBus signals are also not like Qt signals that get removed when the
sender is destroyed. They're match rules that stay in place and can
just be set up once.